Don't touch whitespace on the current line.
It's fairly common to start a method definition and then save
just before starting in on that method. The result is that the
method indentation is removed. This happens in a number of
situations. With this change, whitespace on the current line
is ignored if there's ONLY whitespace on that line.