Switch to using backwardsScan.
This fixes a bug where the cursor index would change out from under
you as the replacements in the scan were happening. This was causing
the cursor row to be mis-identified, which was resulting in inconsistent
row identification (which looked sort of like an off-by-one error).
By scanning from the end of the buffer, we prevent the index from getting
mangled like that and end up with the correct cursor row.